Reorganize code
- Put godeps file handling stuff all in the same place.
- diff outputs name of Godeps/Godeps.json file.
- don't export stuff (lower case function / method names).
