+++ diff.go
@@ -20,9 +20,7 @@ previous 'go save' execution.
 }
 
 func runDiff(cmd *Command, args []string) {
-	var gold Godeps
-
-	_, err := readOldGodeps(&gold)
+	gold, err := loadDefaultGodepsFile()
 	if err != nil {
 		log.Fatalln(err)
 	}
@@ -43,7 +41,7 @@ func runDiff(cmd *Command, args []string) {
 		GoVersion:  ver,
 	}
 
-	err = gnew.Load(dot, dot[0].ImportPath)
+	err = gnew.fill(dot, dot[0].ImportPath)
 	if err != nil {
 		log.Fatalln(err)
 	}
@@ -59,12 +57,12 @@ func runDiff(cmd *Command, args []string) {
 func diffStr(a, b *Godeps) (string, error) {
 	var ab, bb bytes.Buffer
 
-	_, err := a.WriteTo(&ab)
+	_, err := a.writeTo(&ab)
 	if err != nil {
 		log.Fatalln(err)
 	}
 
-	_, err = b.WriteTo(&bb)
+	_, err = b.writeTo(&bb)
 	if err != nil {
 		log.Fatalln(err)
 	}
@@ -72,7 +70,7 @@ func diffStr(a, b *Godeps) (string, error) {
 	diff := difflib.UnifiedDiff{
 		A:        difflib.SplitLines(ab.String()),
 		B:        difflib.SplitLines(bb.String()),
-		FromFile: "Godeps",
+		FromFile: b.file(),
 		ToFile:   "$GOPATH",
 		Context:  10,
 	}
